pyproject.toml
[project]
name = "langchain-examples-for-beginners"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "langchain-pinecone>=0.2.12",
    "langchain-tavily>=0.2.11",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
]
